+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/miam_core/data/datasource/MiamAPIDatasource.kt
@@ -220,7 +220,7 @@ class MiamAPIDatasource : RecipeDataSource, GroceriesListDataSource, PointOfSale
     ): List<Recipe> {
         LogHandler.info("[Miam][MiamAPIDatasource] starting getRecipeSuggestions $criteria")
         val url =
-            "${HttpRoutes.RECIPE_SUGGESTIONS}?supplier_id=${supplierId}&${includedToString(included)}"
+            "${HttpRoutes.RECIPE_SUGGESTIONS}?supplier_id=${supplierId}&page[size]=${size}&${includedToString(included)}"
         val returnValue = this.post<RecordWrapper>(url, criteria)!!.toRecords()
         LogHandler.info("[Miam][MiamAPIDatasource] end getRecipeSuggestions $criteria $returnValue")
         return returnValue.map { record -> record as Recipe }